fix: Update version to 0.1.3-beta
This commit updates the version number in the \_\_init__.py file from 0.1.2-beta to 0.1.3-beta. The reason for this change is to reflect a possible bug fix or improvement made in the Python project. The version update is a minor change and does not affect the functionality of the project.